Migrate Application Notice View from XML to Compose

+++ domain/src/main/java/daily/dayo/domain/model/Notice.kt
@@ -3,7 +3,7 @@ package daily.dayo.domain.model
 import java.io.Serializable
 
 data class Notice(
-    val noticeId: Int,
+    val noticeId: Long,
     val title: String,
     val uploadDate: String
 ) : Serializable
